python-flask/requirements.txt
flask==2.3.3
flask-restful==0.3.10
pyOpenSSL==23.2.0
flask-oauthlib==0.9.6
pyjwt==2.6.0
flask-sqlalchemy==3.0.5
bcrypt==4.0.1